#61 Updated api.py: Added dictionary return for Full history (#70)
* updated api.py

Added dictionary type response whenever user clicks on get full history.
Earlier a list type was returned which was of no use for non developers. Will implement downloading csv option through this dict now

* Updated api.py

get full_history now returns a dictionary format rather that list of list which it used to return earlier. This dict can be used in javascript to convert and download a csv on the go.